Refactor api store
