+++ components/api/table.vue
@@ -1,16 +1,16 @@
 <script lang="ts" setup>
   import EditApi from '../modal/editApi.vue';
   import ModalConfirm from '../modal/confirm.vue';
-  import useApiStore from '~~/stores/useApiStore';
+  import useApi from '~~/stores/useApi';
   import useResourceModelStore from '~~/stores/useResourceModelStore';
   import { Api } from '~~/types/models';
 
   const deps = reactive({ target: '' });
 
-  const api = useApiStore();
+  const api = useApi();
   const resourceModel = useResourceModelStore();
 
-  useMountedFetch([api, resourceModel]);
+  useMountedFetch([resourceModel]);
 
   const editApiModal = useModal(EditApi, {
     id: 'edit-api',
@@ -27,7 +27,7 @@
 
       closeModal();
 
-      await api.fetch({ mutateCache: true });
+      await api.refresh();
 
       deps.target = '';
     },
