+++ stores/index.ts
@@ -1,7 +1,7 @@
-export { default as useApiStore } from './useApiStore';
 export { default as useProjectKeyStore } from './useProjectKeyStore';
 export { default as useResourceDataStore } from './useResourceDataStore';
 export { default as useResourceDataTypeStore } from './useResourceDataTypeStore';
 export { default as useResourceModelStore } from './useResourceModelStore';
 
+export { default as useApi } from './useApi';
 export { default as useProject } from './useProject';
